pyproject.toml
[tool.poetry]
name = "audo-eq"
version = "0.1.0"
description = "A modular, reference-based auto-mastering engine."
authors = ["Skyler Saville <skylersaville@gmail.com>"]
readme = "README.md"

[[tool.poetry.packages]]
include = "audo_eq"
from = "src"

[tool.poetry.scripts]
audo_eq = "audo_eq.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
pedalboard = "^0.7.0"    # Core audio processing
pyloudnorm = "^0.1.0"    # Professional LUFS measurement (critical)
numpy = "^1.24.0"
scipy = "^1.10.0"        # Wavelets, advanced FFT, filters
librosa = "^0.10.0"      # General audio analysis (BPM, features)
pydantic = "^2.0.0"      # Robust configuration & settings validation
soundfile = "^0.12.0"    # Reliable audio I/O (WAV)
pyyaml = "^6.0.1"        # YAML config support


[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
jupyter = "^1.0.0"
black = "^23.0.0"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
